Add support for Python 3.12 (#1359)
### Summary

Now that Aer supports Python 3.12 starting with 0.13.2
(https://github.com/Qiskit/qiskit-aer/pull/2022), we should be able to
support it as well. The only code change needed was to update the
deprecated usage of `datetime.datetime.utcnow()` to
`datetime.datetime.now(datetime.UTC)`. I added a python version checker
to handle this since `datetime.UTC` was only introduced in 3.11.

---------

Co-authored-by: Will Shanks <willshanks@us.ibm.com>