Switch tests and documentation to use qiskit-ibm-runtime fake backends
The IBM-specific fake backends in Qiskit are being removed in Qiskit 1.0
in https://github.com/Qiskit/qiskit/pull/11376. The same fake backends
have been moved to qiskit-ibm-runtime in
https://github.com/Qiskit/qiskit-ibm-runtime/pull/1140 and follow up
PRs.

Here the `qiskit` imports are replaced with `qiskit-ibm-runtime` imports
in order to provide a smooth transition between Qiskit 0.45 and Qiskit
1.0. In the future, it may be preferable to use generic fake backends
maintained in the repo and not depend on `qiskit-ibm-runtime` for them.

Additionally, this commit reorders the requirements-dev.txt file so it
is easier to read and adds qiskit-ibm-runtime as a development
dependency.