Deprecate BackendData.is_simulator (#1390)
No code in qiskit-experiments depends on `is_simulator` already and
`BackendV2` does not provide a standard way to check if a backend uses a
simulator. The code currently tries checking if a backend is a subclass
of some known fake backend classes from qiskit and qiskit-ibm-runtime,
but this code is awkward and seems to be doing something awkward for
little benefit.