Enable loading experiments using the runtime service (#1386)
### Summary

Follow up to #1371 . `ExperimentData.load()` currently requires the
deprecated `qiskit-ibm-provider` provider to load an experiment. This PR
allows the provider parameter to take in the `QiskitRuntimeService`
service as well. Unfortunately, this is confusing since the experiment
service is another parameter for `load()`. We should probably drop the
experiment service parameter in a future release and only accept the
`QiskitRuntimeService` service.

---------

Co-authored-by: Will Shanks <willshanks@us.ibm.com>