Refactor CurveAnalysis tutorial description of ScatterTable
This change refactors the description of `ScatterTable` in the
`CurveAnalysis` tutorial to try to provide a more intuitive picture of
what series and category are. It tries to the split the discussion into
two parts: first a high level description of what the terms are intended
to mean and then a walkthrough of the default `CurveAnalysis` flow
pointing out how the terms are used in practice.

The short version is that `series` and `category` could be used as arbitrary
labels, but in practice `category` is used to distinguish stages of data
processing ("raw", "formatted", "fitted") while `series` is used to
label data associated with a specific function (a model but not
necessarily one written out as a fit model for the class). Together
`series` and `category` (and `analysis` if considering a
`CompositeCurveAnalysis` subclass) should distinguish a specific y vs x
curve.