Add option to plot residuals (#1382)
### Summary

This PR will give the user the option to pass `plot_residual=True` to
the analysis. This will add a residual plot to the figure.
### Details and comments

Some details that should be in this section include:

- Came from an issue that was open #1169 .
- What tests and documentation have been added/updated
- What do users and developers need to know about this change

What needed to do in this PR:
- [x] Add support for different subplot sizes.
- [x] Add automatic residual calculation.
- [x] Plot residuals.
- [x] Style the plot's color, legend, limits, etc.


When setting `plot_residuals=True` in analysis option, a residual plot
will be added. currently work only for experiments with 1 plot in its
figure.
Example for output with `plot_residuals=True` for `RamseyXY` experiment:

![image](https://github.com/Qiskit-Extensions/qiskit-experiments/assets/51112651/206d7900-d196-474f-a87a-4dea0d15c531)

---------

Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>