Remove FakeBackendV2 import deprecated in Qiskit 0.46 (#1420)
### Summary

In qiskit 0.46, several classes in the `qiskit.providers.fake_provider`
module were deprecated, including
`qiskit.providers.fake_provider.fake_backend.FakeBackendV2`. However,
the removal PR failed to address this particular class, so while this
particular `FakeBackendV2` was no longer part of the public API, the
import path would still work. This oversight will be addressed in
https://github.com/Qiskit/qiskit/pull/11997, but in order to merge this
PR we must make sure that all the downstream libraries covered by
`qiskit-neko` are up-to-date with the change.

This change adds an alternative import path for this class in
qiskit-experiments. It follows the pattern from other 1.0 import
deprecations. The removal would roll out in the 1.1 release, but it is
indicated as a 1.0 removal because that is when it should have
technically taken place.

### Details and comments

Blocks https://github.com/Qiskit/qiskit/pull/11997.