Populate older release notes (#1437)
### Summary

As described in
https://github.com/Qiskit-Extensions/qiskit-experiments/pull/1412#pullrequestreview-1889072071,
this PR removes the directive for older release notes to prep for the
0.6.1 patch release. Avoiding rebuilding older release notes every time
also removes the requirement for local builds to fetch older stable
branches. Older patch release notes that were missing, such as 0.3.1,
have been repopulated. The reno version has also been updated to the
newest release instead of the previous commit ID.