Fix a bug in EPG computation for devices with ECR gates (#1431)
### Summary
Fixed a bug in EPG (error per gate) computation in `RBAnalysis` where it
fails with a ``TypeError`` for backends with ECR gate as a 2-qubit basis
gate (e.g. IBM Eagle processors).
Fixed #1419 .

### Details and comments
Added the standard gate error ratio (`1.0`) for ECR gate as suggested in
the issue. Also added a sentence to explain how to skip EPG computation
in the manual.