Update version string in stable docs deploy CI (#1462)
This PR replaces the buggy `git describe` command, which doesn't
register the updated version when doing a release, with the version in
`VERSION.txt`. We should test that this works by re-running the stable
docs deploy on `main` afterwards. Edit: maybe not since this would build
to `stable/0.8` right now after the version bump.