Deprecate pulse and restless related experiments and classes
This change deprecates the experiments that rely on scanning the
parameters of pulses in pulse gate calibrations. Qiskit 2.0 will remove
support for pulse gate calibrations, making these experiments impossible
to run. The `Calibrations` and `BasisGateLibrary` classes are also
deprecated since they have no use without pulse gate calibrations to
track. It is planned that Qiskit Pulse will be moved to Qiskit Dynamics
and perhaps the experiments and calibrations can be adapted to that use
case for calibrating simulated experiments. For now though, this code is
removed from Qiskit Experiments to help with making the package more
maintainable. Some additional helper code (like the experiments analysis
classes and methods of `BackendTiming` and `BackendData`) are also
deprecated.

Support for restless experiments is also deprecated with this change.
Restless support is distinct from pulse support, but it is deprecated
with the same motivation of simplifying the package overall. With
improvements in the reliability of IBM Quantum's qubit initialization,
circuit exectuion has already become reasonably fast and restless
measurements do not add much performance improvement. It is expected
that the restless features are little used as there has been no user
feedback about them.