Address `c_if`, `Permutation`, and `transpile` parameter deprecations in Qiskit 1.3 (#1482)
* `Permutation` instances in tomography experiments have been replaced
with `PermutationGate` instances that work the same way for those
experiments. The only user facing change should be the avoidance of
deprecation warnings in Qiskit 1.3.
* `c_if` usage on conditional gates in some tests has been replaced with
`if_test` blocks around those gates.
* `ParallelExperiment` now considers a `target` transpiler option when
determining how large a circuit to make when combining subexperiments
into a larger circuit. Previously it considered the number of qubits in
a `coupling_map` transpiler option or in the experiment's backend.
* Use of `instruction_durations` as a transpile option was replaced with
`target` in one test (`instruction_durations` was deprecated as a
transpiler parameter).
* Add explicit `order="F"` argument to `cvxpy.vec()` calls. This
function started issuing a `FutureWarning` in the latest version of
cvxpy about the default order changing from F to C.