Remove remaining references to Provider classes
This chane is a follow up to 519bb534c134ca9a251f4f92f6abd39f56017341 to
remove remaining references to the Provider base classes which were
deprecated in Qiskit 1.1. Providers are now unspecified objects that
generate backends. As a follow up, work is needed on the
`ExperimentData` class to tighten up how it works with providers,
especially since the main provider now is `QiskitRuntimeService` from
qiskit-ibm-runtime, which already was not actually a `Provider`.