Remove workarounds for passing run options to sampler
This runtime patching code was added along with the initial support for
sampler execution because the path through
`qiskit_ibm_runtime.SamplerV2` through to `BackendSamplerV2` did not
carry through options for level 1 measurements or noise models. This
path is used in the tests (while execution on physical backends with
`SamplerV2` does not use this path). Now with Qiskit 1.3 and
qiskit-ibm-runtime 0.34, this patching is no longer needed. Remaining
warnings filtering was moved to the calibrations tutorial where it was
needed.