Fix tests failing against Qiskit main due to change in delay behavior
A long-standing bug in qpy leads to the units of delay instructions
being overwritten to `dt`. See:

https://github.com/Qiskit/qiskit/issues/10662

Some experiments test that their circuits are round-trip serializable
through qpy (perhaps not as necessary as it once was as the room for
sticking arbitrary Python into circuits is reduced with the move to
Rust). The T1 and Tphi circuits have delays and so were affected by the
qpy bug. However, there is a second bug related to delays which canceled
this bug out for those tests. `Delay.__eq__` does not check that the
units are the same for the two delay instructions, so the loss of the
original unit did not matter for the tests. Following this change:

https://github.com/Qiskit/qiskit/pull/13486

Comparing circuits with delays now includes the delay units in the
determination of equality. `Delay.__eq__` itself still does not consider
the unit but some part of adding the delay to a circuit converts it into
a Rust object that does check the unit.

Here the tests are given a backend with a `dt` value so that the
experiments will generate circuits with delays in `dt` and so still have
delays with the same `dt` after the qpy roundtrip.