Replace backend comparison test with comparison of backend attributes
Starting with qiskit-ibm-runtime 0.35, running circuits with
`qiskit_ibm_runtime.SamplerV2` produces jobs whose `backend()`
method returns a deepcopy of the original backend object passed to the
sampler. Since `BackendV2` does not override `__eq__`, it does not
compare equal to a copy of itself and the test that checks that the
original backend equals the backend in the experiment data started to
fail.

Here the test code is updated to check that the main properties of the
backend compare equal rather than comparing the backend objects
directly.