Replace setup.py with pyproject.toml configuration
qiskit-experiments is a simple pure Python project and all of its
packaging configuration can be specified statically with pyproject.toml.
The build configuration should be the same as it was previously.  The
requirements files and version file are left external pyproject.toml as
they were external to setup.py previously. The build backend is left as
setuptools.