support RuntimeJobV2 job status (#1512)
The job status definition in `RuntimeJobV2`
([JobStatus](https://github.com/Qiskit/qiskit-ibm-runtime/blob/main/qiskit_ibm_runtime/runtime_job_v2.py#L42-L48))
uses a str object compare to the previous definition in `RuntimeJob`
which is an [enum
](https://github.com/Qiskit/qiskit/blob/stable/1.3/qiskit/providers/jobstatus.py#L18-L27).
This cause qiskit-experiment to report `JobStatus.DONE` immediately
after running an experiment, even if the job was queue/running and a
wrong `ExperimentStatus`.
This change fixes this issue by supporting both definitions, it converts
the str `RuntimeJobV2` to the original enum (with the same names).

---------

Co-authored-by: Will Shanks <willshanks@us.ibm.com>