Remove pulse and calibration management related code (#1511)
This change prepares the package for Qiskit 2.0 by removing all features
related to Qiskit Pulse which is removed in Qiskit 2.0 (see for example
[Qiskit#13872](https://github.com/Qiskit/qiskit/pull/13872)). It
includes several related changes:

* Remove all experiments that require pulse schedules to run on a
backend with a set of standard basis gates. The set of removed
experiments includes `Rabi` and `QubitSpectroscopy` for example.
Additionally, analysis classes used only the removed experiments are
also removed. These classes should have all emitted deprecation warnings
in the previous release.
* Remove all code related pulse calibration management, including the
`Calibrations` and `BasisGateLibrary` classes and the calibration
experiment classes. These classes should have all emitted deprecation
warnings in the previous release.
* Deprecate `PulseBackend` and drop `qiskit-dynamics` as a dependency of
`qiskit-experiments[extras]`. This is a last dependency on Qiskit Pulse
that is not fully dropped yet because it was not previously deprecated.
* Convert the `jupyter-execute` cells in the docs to use Qiskit Aer or
mock backends instead of using `PulseBackend`.
* Remove pulse support from other experiments and classes that have not
been removed including methods of `BackendData` and `BackendTiming` like
`pulse_time` and the feature of `MultiStateDiscrimination` that allowed
it to accept pulse schedules for exciting higher qubit states. Also,
remove the ability to serialize and deserialize pulse schedule blocks
with `ExperimentEncoder` and `ExperimentDecoder`.
* Deprecate `FineZXAmplitude` as it can not be easily used without a
pulse calibration to implement an RZX rotation.
* Update examples in the documentation that previously used removed
experiments to use other experiments that have not been removed.
* Remove support for custom pulse calibrations from randomized
benchmarking experiments and `ParallelExperiment`.
* Add a private `ZZRamseyTestBackend` for the `ZZRamsey` tests and
documentation that is similar to `T2HahnBackend` to provide more
consistency with the other experiment examples that were updated not to
use `PulseBackend`.

Note that a few small pulse features might have been removed without a
release cycle of deprecation warnings. Since the IBM Quantum service was
the main use case of the Qiskit Pulse features and it has already
stopped accepting jobs with Qiskit Pulse calibrations, it was deemed
better to do a clean removal of all pulse features than to adhere
strictly to the deprecation policy.

The changes here focus on the Qiskit Pulse related updates needed for
compatibility with Qiskit 2.0. Some further follow up will also be
necessary to address other incompatibilities with Qiskit 2.0.