Remove RestlessMixin and deprecate restless nodes (#1524)
This change removes the previously deprecated `RestlessMixin` class and
deprecates the restless processing nodes `RestlessToIQ` and
`RestlessToCounts`. The motivation for the removals is to streamline the
remaining code base. The restless features are little used and it
remains possible to set up experiments to run in restless mode by using
a custom data processor even without support in the base package.