Replace setup.py with pyproject.toml configuration (#1515)
qiskit-experiments is a simple pure Python project and all of its
packaging configuration can be specified statically with pyproject.toml.
The build configuration should be the same as it was previously. The
build backend is left as setuptools.

Detailed changes:

* Consolidate dependencies, optional dependencies, and development
dependencies into pyproject.toml (with extras and dependency groups)
* Add `setuptools_scm` as a build dependency to ensure package files are
always included unless opted out of (and opt out of the docs and CI
files but keep the tests).
* Update tox.ini to use some newer conventions (`package` instead of
`use_develop`, set `PYTHONSAFEPATH`, use tox's default install command).