Remove nbsphinx gallery workaround
The nbsphinx gallery css file seems to be added properly now. The
workaround of vendoring it was causing a conflict with the main nbsphinx
file, leading sphinx to issue a warning on a second build of the docs,
so this issue did not affect continuous integration where the docs are
always built fresh but did affect local builds.