Use pypa/build and tox-uv in CI testing and publishing
pypa/build builds an sdist and then a wheel from the sdist which helps
avoid leaving files out of the sdist. Additionally, pypa/build formats
the package file names in the way that PyPI will soon require while
`python setup.py` does not.

Setting up the .pkg environment led to tox errors when `package` was set
to different values in different environments so they were all set to
editable by default but they can be overridden with `--override
testenv.package=wheel`.

Additionally, the shared `.tox/lint` `envdir` in `tox.ini` was dropped
because with `tox` version 4 it resulted in the environment being
rebuilt every time a different command sharing that environment
directory was run (with `tox` 3 the environment was reused). Instead, a
recommendation to use `tox-uv` was added to the documentation. `uv` is
quick at creating similar virtual environments so it allows using many
`tox` environments without worrying about speed or disk space (because
it uses hard links).

Explicit installation of `pyarrow` was removed from `tox.ini` because
Pandas no longer issues a deprecation warning when `pyarrow` is not
installed.

Along with suggesting using `tox-uv` in the documentation, the GitHub
workflows were updated to use `uv` and `tox-uv`.

Also, the version of Python used for running linting and building the
release artifacts was updated from 3.9 to 3.12

Closes https://github.com/qiskit-community/qiskit-experiments/issues/1527