Prepare for Qiskit 2.0 by addressing removed classes and methods (#1526)
This change prepares for Qiskit 2.0 by guarding or replacing references
to objects removed in Qiskit 2.0.

## Detailed list of changes:

### Compatibility changes:

* Guard references to `BackendV1`. `BackendV1` was removed in Qiskit
2.0. There were only a few remaining references to `BackendV1` after the
previous removal of pulse support, so it was not hard to keep the
remaining support for now but it should be removed when Qiskit 1 is no
longer supported.
* Update usage of `qiskit.result.Result.header` to expect a dict (Qiskit
2) as well as a class (Qiskit 1). `ExperimentData` gets the circuit
metadata from the header (by convention the metadata is copied into the
result and `ExperimentData` relies on this).
* Add classes `MeasLevel` and `MeasReturnType` to enumerate the expected
measurement level (kerneled or classified) and return type (averaged or
not). Previously, private classes of Qiskit were inadvertently used for
these purposes but they were removed in Qiskit 2.0.
* Remove null default `inst_map` transpile options from experiments.
These should have been removed along with the rest of pulse support.
* Update tests to use `GenericBackendV2` or a custom `BackendV2` in
place of removed `FakeOpenPulse2Q` and `Fake5QV1` backends from Qiskit.
* Remove suppression of Qiskit deprecation warnings from the tests.
* Remove stray references to `ProviderV1` in tests.
* Update the saved `QuantumVolume` results for Qiskit 2.0's change in
the quantum volume algorithm.
* Vendor deprecation decorators to prepare for their removal from
Qiskit's public API (see
[here](https://github.com/Qiskit/qiskit/pull/11296#pullrequestreview-2665006544)).
At first it appeared that these methods would be changed late in the
Qiskit 2.0 cycle but the change was pushed back in time. Still the plan
is to make the functions private, so we need to stop using them.

### Other changes

* Provide a more specific warning in `LocalReadouMitigator` if it is
used with a backend without a ``properties()`` method or with a backend
that does not the required readout error properties.
* Fix a performance issue in `QuantumVolume`. When Qiskit Aer is
installed, `QuantumVolume` uses `AerSimulator` to simulate the outcomes
of the quantum volume circuits to determine the heavy output
probabilities. In its default instantiation, `AerSimulator` regenerates
its `Target` every time it is accessed and this generation takes an
appreciable amount of time. Because the target was being accessed
separately for every circuit, this overhead could accumulate to over a
minute in the standard 100 circuit experiment. This overhead was reduced
by processing all the circuits in one pass.
* Switch `FineFrequency` tests to use `T2HahnBackend` in place
`MockIQBackend` since it provides a more realistic backend emulation.
* Fix the `stacklevel` on the deprecation warning for `RestlessNode`.

---------

Co-authored-by: Eli Arbel <arbel@il.ibm.com>