Rework nbsphinx gallery workaround (#1528)
The Sphinx config had a workaround to copy a vendored version of the
nbsphinx gallery css file and add it to the list of custom styles. This
workaround worked in getting the css file included in the html output
but it led to a conflict with nbsphinx which was also adding this css
file to the `_static` directory.

Sphinx would raise a warning when nbsphinx would try to overwrite the
custom-style version of the css file. nbsphinx adds its file first and
there is no error for the custom style overwriting it, so in a clean
directory there was no error, but on second run Sphinx would issue a
warning about the overwrite. The vendored css file was susceptible to
falling out of sync with updates to nbsphinx.

Here, we instead call `app.add_css_file('nbsphinx-gallery.css')`
explicitly which does get the file added to the html output but leaves
the copying of the css file to `_static` to nbsphinx. There is probably
a bug that could be filed upstream here. It is likely not a common case
to use `nbgallery` without any other nbsphinx features.

Along with the fix to the gallery css file, the gallery thumbnail for
readout mitigation was also fixed. It seems like the file names are
based on the jupyter cells of the manuals and can change if the content
of the manual is modified (ideally Sphinx would warn when an image is
missing; or we would find a way to save the images with stable names).