Split development dependencies into groups (#1534)
Here the development dependencies in the `dev` dependency group are
split into subgroups based on purpose. The `dev` group still encompasses
all of the development dependencies through `include-group`. The primary
benefit of this change is that `tox run -eblack` now only needs to
install `black` instead of a full development environment, while the
version of `black` remains pinned in a single place in the repo. There
may be some marginal benefit to avoiding installing sphinx in the test
environment or stestr in the docs environment as well.

Additionally, this change makes some small changes to the development
dependencies, removing a pin on the Sphinx version (added because Sphinx
extensions had problems with a new Sphinx release) and `pyarrow` (added
because Pandas was issuing warnings when it was not installed but has
since removed the warning).